# docs/swagger.yaml
definitions:
...
  episode.Episode:
    properties:
      description:
        type: string
      duration_minutes:
        type: integer
      episode_number:
        type: integer
      id:
        type: integer
      name:
        type: string
      release_date:
        type: string
      season_id:
        type: integer
      url:
        type: string
      user_id:
        type: integer
    required:
    - duration_minutes
    - episode_number
    - name
    - release_date
    - url
    type: object
...
  routes.CategoryRequest:
    properties:
      category_name:
        type: string
      description:
        type: string
    required:
    - category_name
    - description
    type: object
...
  routes.DefaultMessageResponse:
    properties:
      message:
        type: string
    type: object
  routes.EpisodeRequest:
    properties:
      description:
        type: string
      duration_minutes:
        type: integer
      episode_number:
        type: integer
      name:
        type: string
      release_date:
        type: string
      url:
        type: string
    required:
    - duration_minutes
    - episode_number
    - name
    - release_date
    - url
    type: object
...
  routes.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  routes.SignInResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/categories:
    get:
      description: Retrieves all categories
      produces:
      - application/json
      responses:
        "200":
          description: Categories Retrieved
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Categories not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all categories
    post:
      description: Creates a new category
      parameters:
      - description: Category object to be created
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/routes.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category Created
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "409":
          description: Category already exists
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a category
  /api/categories/{id}:
    delete:
      description: Deletes a category with the specified ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category Deleted
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Cannot delete category because it is used in movies
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a category
    get:
      description: Retrieves the category with the specified ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category Retrieved
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a category
    put:
      description: Updates an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated category object
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/routes.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category Updated
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Category Not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "409":
          description: Category with this name already exists
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a category
...
  /api/check-auth:
    post:
      consumes:
      - application/json
      description: Check user authentication and return user role
      produces:
      - application/json
      responses:
        "200":
          description: Successfully checked authentication
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: 'Unauthorized: User not authenticated'
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Check authentication
      tags:
      - authentication
  /api/episodes/{id}:
    delete:
      description: Deletes an episode with the specified ID
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Episode deleted
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Episode Not Found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete an episode
    get:
      description: Retrieves an episode with the specified ID
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Episode
          schema:
            $ref: '#/definitions/episode.Episode'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Episode not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      summary: Retrieve an episode
    post:
      description: Creates a single episode for the specified season
      parameters:
      - description: Season ID
        in: path
        name: id
        required: true
        type: integer
      - description: Episode
        in: body
        name: episode
        required: true
        schema:
          $ref: '#/definitions/routes.EpisodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Episode added
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Season not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Create episode
    put:
      description: Updates an existing episode
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: integer
      - description: Episode
        in: body
        name: episode
        required: true
        schema:
          $ref: '#/definitions/routes.EpisodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Episode updated
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "404":
          description: Episode not found
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Update episode
...
  /api/signIn:
    post:
      consumes:
      - application/json
      description: 'Sign in with email and password to obtain JWT token. Creditnails
        for sign in "email": "big@example.com",  "password": "Aa12345678#"'
      parameters:
      - description: Email and password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/routes.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully signed in
          schema:
            $ref: '#/definitions/routes.SignInResponse'
        "400":
          description: Invalid data format
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "401":
          description: User does not exist or incorrect password
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/routes.DefaultMessageResponse'
      summary: Sign in
      tags:
      - authentication
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"